Update FinalProject.ipynb
The hash tagged lines of code are for if we would like to tweet out text from a file we created. With a little bit of tinkering we could create a list of tweets that it would send out from the text file every so often (which is why we imported time).